refactor: store less data in local auth cookie
