+++ src/components/pages/auth/login.tsx
@@ -7,8 +7,8 @@ import { useEffect, useState } from 'react'
 
 import { filterUnwantedRedirection } from './utils'
 import { getAuthPayloadFromSession } from '@/auth/auth-provider'
-import { fetchAndPersistAuthSession } from '@/auth/fetch-auth-session'
 import { kratos } from '@/auth/kratos'
+import { fetchAndPersistAuthSession } from '@/auth/local-session/fetch-and-persist-auth-session'
 import type { AxiosError } from '@/auth/types'
 import { useAuth } from '@/auth/use-auth'
 import { useCheckInstance } from '@/auth/use-check-instance'
