+++ src/components/pages/auth/settings.tsx
@@ -5,8 +5,8 @@ import {
 import { useRouter } from 'next/router'
 import { useEffect, useState } from 'react'
 
-import { fetchAndPersistAuthSession } from '@/auth/fetch-auth-session'
 import { kratos } from '@/auth/kratos'
+import { fetchAndPersistAuthSession } from '@/auth/local-session/fetch-and-persist-auth-session'
 import { useAuth } from '@/auth/use-auth'
 import { useCheckInstance } from '@/auth/use-check-instance'
 import { Flow, FlowType, handleFlowError } from '@/components/auth/flow'
